Add a bunch more band3 classes, try to fill in context wherever possible (#373)
* try to fill in missing ctx for overshell classes

* gamemode re-integrated

* gamemode has linkissues

* add gameconfig and psprovider

* modifiermgr work

* expand bandsongmgr and add metaperformer

* add bandsongmetadata

* bandsongmetadata context

* bandsongmetadata getters

* bandsongmetadata save and load

* add bandsongmgr

* inline find

* add clearandshrink inline method

* little more bandsongmetadata work

* bandsongmgr work

+++ src/system/os/DateTime.cpp
@@ -83,7 +83,7 @@ void DateTime::ToMiniDateString(class String& str) const {
     str += MakeString("%02d/%02d", mMonth + 1, mDay);
 }
 
-unsigned short DateTime::Year() const {
+int DateTime::Year() const {
     return mYear + 1900;
 }
